backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
groq
google-generativeai
httpx
PyMuPDF
python-docx
Pillow
chromadb
pydantic
pydantic-settings
sse-starlette
jinja2
langdetect
langchain-text-splitters
openai
firebase-admin
web3
